app/requirements.txt
Flask==3.1.0
Flask-Limiter==3.9.0
Flask-Caching==2.3.0
requests==2.32.3
python-dotenv==1.1.0
pytest==8.3.5
pytest-cov==6.1.0
pytest-mock==3.14.0
responses==0.25.7
